WIP: Readme and beginnings of a script to check for valid UTF8
This initial commit includes the sample CSVs, a reddme that details my perspective on the exercise and the script that I've got going so far. I take the data from stdin and put it in a csv so that it's easier to iterate over as I check if a given string is valid utf-8 or not.