Check for valid utf-8
Taking mostly from a resource I found online (https://thoughtbot.com/blog/fight-back-utf-8-invalid-byte-sequences), the code
now checks if a string is valid utf-8. If it's not then the string is replaced with an empty string. With more time, I'd adjust
this method so that I can print a a warning to `stderr`.